+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - **Breaking Change**: restrict implicit bool coercion for ``FD``, ``T?`` type
     - only allowed in the following conditional context
         - if, for, while, assert
+- **Breaking Change**: change ``Any`` type expression to ``String`` cast semantics.
+    - now do not perform string coercion (to-string)
 - adjust error message layout and colors
 - improve semantic error messages
     - show ``did you mean ?`` suggestions for undefined variable, undefined field/method, undefined type